fix(collector): add socketcall ID based conversion helper
Extract a helper to convert socketcall to network calls based
on the socketcall identifier.  This helper is required for the
enter probe of the custom collector BPF probe; because at that
point the socketcall arguments are not yet stored in the BPF
argument map.

Because of direct-attaching to the socketcall tp, the socketcall
ID is available in the context / tracepoint format/structure.

See also respective custom collector BPF probe update.

This is specific to RHACS / custom collector BPF probe.

Signed-off-by: Hendrik Brueckner <brueckner@de.ibm.com>